Criado as rotas e um header (esboço)
