+++ pokedex/src/App.js
@@ -1,4 +1,3 @@
-import logo from './logo.svg';
 import './App.css';
 
 const App =()=>{
