+++ pokedex/src/components/pages/Error/Error.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Error = () => {
+
+    return(
+        <>
+        <h1>Pagina de erro!</h1>
+        </>
+    )
+}
+
+export default Error;
