--- pokedex/src/components/pages/Home/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-const Home = () => {
-    return(<>Home page</>)
-}
-export default Home;
