+++ pokedex/src/components/pages/Pokedex/PokedexDetails/PokedexDetails.js
@@ -1,0 +1,20 @@
+import React from "react";
+import Header from "../../../shared/Header/Header"
+import styled from "styled-components"
+
+const PokedexDetails = () => {
+
+    const ContainerHeader = styled.div`
+        display: flex;
+        direction: column;
+        align-items: center;
+        justify-content: space-around;
+        `
+
+    return(
+        <ContainerHeader>
+            <Header />
+        </ContainerHeader>
+    )
+}
+export default PokedexDetails;
