--- pokedex/src/components/pages/Pokedex/PokedexDetails/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-const PokedexDetails = () => {
-    return(<>PokedexDetails page</>)
-}
-export default PokedexDetails;
