--- pokedex/src/components/pages/Pokedex/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-const Pokedex = () => {
-    return(<>Pokedex page</>)
-}
-export default Pokedex;
