--- pokedex/src/components/routers/index.js
@@ -1,12 +1,0 @@
-import { BrowserRouter, Route, Routes } from "react-router-dom"
-import Home from "../pages/Home";
-
-const Routers = () => {
-    return(<BrowserRouter>
-        <Routes>
-            <Route path="/" element={<Home />} />
-        </Routes>
-    </BrowserRouter>)
-}
-
-export default Routers;
