+++ pokedex/src/components/shared/Button/Button.js
@@ -1,6 +1,10 @@
 import React from "react";
 
 const Button = () => {
-    return(<>Button page</>)
+    return(
+    <>
+        <button>Botão</button>
+    </>
+    )
 }
 export default Button;
