+++ pokedex/src/components/shared/Header/Header.js
@@ -1,0 +1,19 @@
+import React from "react";
+import styled from "styled-components";
+
+
+const Header = () => {
+
+    return(
+    <>
+        <p>
+            Logo 
+        </p>
+        <h2>Pokemóns</h2>
+        <button>
+            Pokedex/Home
+        </button>
+    </>
+    )
+}
+export default Header;
