--- pokedex/src/components/shared/Header/index.js
@@ -1,6 +1,0 @@
-import React from "react";
-
-const Header = () => {
-    return(<>Header page</>)
-}
-export default Header;
