+++ pokedex/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import Routers from './components/routers';
+import Routers from './components/routers/Routes';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
